refactor(model): rename Read/Write model pipelines to Decode/Encode
Rename the json/avro/protobuf model pipeline classes (and their tests) from
the Read/Write naming to Decode/Encode, aligning with the decoder/encoder
direction the model handler vends and preparing for the upcoming removal of
ValidatorHandler and ConverterHandler.

Refs #1956
