+++ runtime/model-avro/src/main/java/io/aklivity/zilla/runtime/model/avro/internal/AvroModelHandlerImpl.java
@@ -41,8 +41,8 @@ import io.aklivity.zilla.runtime.engine.model.function.ValueConsumer;
 import io.aklivity.zilla.runtime.model.avro.config.AvroModelConfig;
 
 // Per-worker factory for an Avro model. One handler serves both directions: supplyDecoder vends a
-// per-stream AvroReadModelPipeline (catalog framing stripped, value validated and re-encoded) and supplyEncoder
-// vends a per-stream AvroWriteModelPipeline (catalog framing emitted, value validated). Configuration-derived
+// per-stream AvroDecodeModelPipeline (catalog framing stripped, value validated and re-encoded) and supplyEncoder
+// vends a per-stream AvroEncodeModelPipeline (catalog framing emitted, value validated). Configuration-derived
 // state (catalog, schema cache, extraction paths) is shared; in-flight state lives on each pipeline.
 public final class AvroModelHandlerImpl extends AvroModelHandler implements ModelHandler
 {
@@ -83,14 +83,14 @@ public final class AvroModelHandlerImpl extends AvroModelHandler implements Mode
     public ModelPipeline supplyDecoder(
         ModelVisitor visitor)
     {
-        return new AvroReadModelPipeline(this, paths, names, visitor);
+        return new AvroDecodeModelPipeline(this, paths, names, visitor);
     }
 
     @Override
     public ModelPipeline supplyEncoder(
         ModelVisitor visitor)
     {
-        return new AvroWriteModelPipeline(this);
+        return new AvroEncodeModelPipeline(this);
     }
 
     int decodePadding(
