+++ runtime/model-avro/src/test/java/io/aklivity/zilla/runtime/model/avro/internal/AvroEncodeModelPipelineTest.java
@@ -12,11 +12,12 @@
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.aklivity.zilla.runtime.model.protobuf.internal;
+package io.aklivity.zilla.runtime.model.avro.internal;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -28,6 +29,7 @@ import org.agrona.concurrent.UnsafeBuffer;
 import org.junit.Before;
 import org.junit.Test;
 
+import io.aklivity.zilla.runtime.engine.Configuration;
 import io.aklivity.zilla.runtime.engine.EngineContext;
 import io.aklivity.zilla.runtime.engine.binding.function.MessageConsumer;
 import io.aklivity.zilla.runtime.engine.config.CatalogConfig;
@@ -38,36 +40,32 @@ import io.aklivity.zilla.runtime.engine.model.ModelVisitor;
 import io.aklivity.zilla.runtime.engine.test.internal.catalog.TestCatalogHandler;
 import io.aklivity.zilla.runtime.engine.test.internal.catalog.config.TestCatalogConfig;
 import io.aklivity.zilla.runtime.engine.test.internal.catalog.config.TestCatalogOptionsConfig;
-import io.aklivity.zilla.runtime.model.protobuf.config.ProtobufModelConfig;
+import io.aklivity.zilla.runtime.model.avro.config.AvroModelConfig;
 
-public class ProtobufWriteModelPipelineTest
+public class AvroEncodeModelPipelineTest
 {
-    private static final String SCHEMA = """
-                                            syntax = "proto3";
-                                            package io.aklivity.examples.clients.proto;
-                                            message SimpleMessage {
-                                                string content = 1;
-                                                optional string date_time = 2;
-                                            }
-                                            """;
+    private static final String SCHEMA = "{\"fields\":[{\"name\":\"id\",\"type\":\"string\"}," +
+        "{\"name\":\"status\",\"type\":\"string\"}]," +
+        "\"name\":\"Event\",\"namespace\":\"io.aklivity.example\",\"type\":\"record\"}";
 
-    private static final String JSON = "{\"content\":\"OK\",\"date_time\":\"01012024\"}";
-    // single zero index byte for the first top-level message, then content="OK" and date_time="01012024"
-    private static final byte[] WIRE =
-        {0x00, 0x0a, 0x02, 0x4f, 0x4b, 0x12, 0x08, 0x30, 0x31, 0x30, 0x31, 0x32, 0x30, 0x32, 0x34};
+    private static final String JSON = "{\"id\":\"id0\",\"status\":\"positive\"}";
+    // id="id0" (len 3) then status="positive" (len 8); the TestCatalog adds no framing prefix
+    private static final byte[] AVRO = {0x06, 0x69, 0x64, 0x30, 0x10, 0x70, 0x6f, 0x73, 0x69, 0x74, 0x69, 0x76, 0x65};
 
     private EngineContext context;
+    private AvroModelConfiguration config;
 
     @Before
     public void init()
     {
+        config = new AvroModelConfiguration(new Configuration());
         context = mock(EngineContext.class);
     }
 
     @Test
     public void shouldTransformWholeValue()
     {
-        ProtobufModelHandlerImpl handler = newHandler();
+        AvroModelHandlerImpl handler = newHandler();
         ModelPipeline pipeline = handler.supplyEncoder(ModelVisitor.NONE);
 
         byte[] in = JSON.getBytes(UTF_8);
@@ -79,28 +77,19 @@ public class ProtobufWriteModelPipelineTest
         assertEquals(in.length, result.consumed());
         byte[] out = new byte[result.produced()];
         dst.getBytes(0, out);
-        assertArrayEquals(WIRE, out);
-
-        // reset and reuse the same pipeline for the next value
-        pipeline.reset();
-        ModelPipelineResult reused = pipeline.transform(0L, 0L, ModelPipeline.FLAGS_COMPLETE,
-            new UnsafeBuffer(in), 0, in.length, dst, 0, dst.capacity());
-        assertEquals(ModelStatus.COMPLETE, reused.status());
-        byte[] outReused = new byte[reused.produced()];
-        dst.getBytes(0, outReused);
-        assertArrayEquals(WIRE, outReused);
+        assertArrayEquals(AVRO, out);
     }
 
     @Test
     public void shouldIsolateInterleavedStreams()
     {
-        ProtobufModelHandlerImpl handler = newHandler();
+        AvroModelHandlerImpl handler = newHandler();
         // two per-stream pipelines from the same per-worker handler
         ModelPipeline a = handler.supplyEncoder(ModelVisitor.NONE);
         ModelPipeline b = handler.supplyEncoder(ModelVisitor.NONE);
 
-        byte[] a1 = "{\"content\":\"OK\",".getBytes(UTF_8);
-        byte[] a2tail = "\"date_time\":\"01012024\"}".getBytes(UTF_8);
+        byte[] a1 = "{\"id\":\"id0\",".getBytes(UTF_8);
+        byte[] a2tail = "\"status\":\"positive\"}".getBytes(UTF_8);
         MutableDirectBuffer dst = new UnsafeBuffer(new byte[256]);
         ByteArrayOutputStream outA = new ByteArrayOutputStream();
 
@@ -117,7 +106,7 @@ public class ProtobufWriteModelPipelineTest
         assertEquals(ModelStatus.COMPLETE, rb.status());
         byte[] outB = new byte[rb.produced()];
         dst.getBytes(0, outB);
-        assertArrayEquals(WIRE, outB);
+        assertArrayEquals(AVRO, outB);
 
         // stream A: finish, prepending A's unconsumed remainder (the caller's decode-slot residue)
         byte[] a2 = concat(a1, ra1.consumed(), a2tail);
@@ -126,54 +115,7 @@ public class ProtobufWriteModelPipelineTest
         assertEquals(ModelStatus.COMPLETE, ra2.status());
         drain(dst, ra2.produced(), outA);
 
-        assertArrayEquals(WIRE, outA.toByteArray());
-    }
-
-    @Test
-    public void shouldDrainOnOverflow()
-    {
-        ProtobufModelHandlerImpl handler = newHandler();
-        ModelPipeline pipeline = handler.supplyEncoder(ModelVisitor.NONE);
-
-        // a 2000-byte content value forces the wire output past a small destination window, exercising the
-        // bounded-chunk OVERFLOW drain across re-transforms (INIT cleared on every re-call after the first)
-        String content = "A".repeat(2000);
-        byte[] in = ("{\"content\":\"" + content + "\",\"date_time\":\"01012024\"}").getBytes(UTF_8);
-        MutableDirectBuffer dst = new UnsafeBuffer(new byte[512]);
-        ByteArrayOutputStream out = new ByteArrayOutputStream();
-        int flags = ModelPipeline.FLAGS_COMPLETE;
-        ModelPipelineResult result;
-        int guard = 0;
-        do
-        {
-            result = pipeline.transform(0L, 0L, flags, new UnsafeBuffer(in), 0, in.length, dst, 0, dst.capacity());
-            drain(dst, result.produced(), out);
-            flags = ModelPipeline.FLAGS_FIN;
-            guard++;
-        }
-        while (result.status() == ModelStatus.OVERFLOW && guard < 1000);
-
-        assertEquals(ModelStatus.COMPLETE, result.status());
-        byte[] wire = out.toByteArray();
-        // the single zero index byte then a content field longer than the destination window
-        assertEquals(0x00, wire[0]);
-        org.junit.Assert.assertTrue(wire.length > 2000);
-    }
-
-    @Test
-    public void shouldRejectUnknownRecord()
-    {
-        when(context.clock()).thenReturn(Clock.systemUTC());
-        when(context.supplyEventWriter()).thenReturn(mock(MessageConsumer.class));
-        ProtobufModelHandlerImpl handler = newHandler("Nonexistent");
-        ModelPipeline pipeline = handler.supplyEncoder(ModelVisitor.NONE);
-
-        byte[] in = JSON.getBytes(UTF_8);
-        MutableDirectBuffer dst = new UnsafeBuffer(new byte[256]);
-        ModelPipelineResult result = pipeline.transform(0L, 0L, ModelPipeline.FLAGS_COMPLETE,
-            new UnsafeBuffer(in), 0, in.length, dst, 0, dst.capacity());
-
-        assertEquals(ModelStatus.REJECTED, result.status());
+        assertArrayEquals(AVRO, outA.toByteArray());
     }
 
     @Test
@@ -181,11 +123,11 @@ public class ProtobufWriteModelPipelineTest
     {
         when(context.clock()).thenReturn(Clock.systemUTC());
         when(context.supplyEventWriter()).thenReturn(mock(MessageConsumer.class));
-        ProtobufModelHandlerImpl handler = newHandler();
+        AvroModelHandlerImpl handler = newHandler();
         ModelPipeline pipeline = handler.supplyEncoder(ModelVisitor.NONE);
 
-        // an unknown field is rejected by the strict JSON parser
-        byte[] in = "{\"content\":\"OK\",\"unexpected\":\"value\"}".getBytes(UTF_8);
+        // id must be a string, not a number
+        byte[] in = "{\"id\":123,\"status\":\"positive\"}".getBytes(UTF_8);
         MutableDirectBuffer dst = new UnsafeBuffer(new byte[256]);
         ModelPipelineResult result = pipeline.transform(0L, 0L, ModelPipeline.FLAGS_COMPLETE,
             new UnsafeBuffer(in), 0, in.length, dst, 0, dst.capacity());
@@ -193,37 +135,40 @@ public class ProtobufWriteModelPipelineTest
         assertEquals(ModelStatus.REJECTED, result.status());
     }
 
-    private ProtobufModelHandlerImpl newHandler()
+    @Test
+    public void shouldReportEncodePadding()
     {
-        return newHandler("SimpleMessage");
+        AvroModelHandlerImpl handler = newHandler();
+        ModelPipeline pipeline = handler.supplyEncoder(ModelVisitor.NONE);
+
+        byte[] in = JSON.getBytes(UTF_8);
+        assertTrue(pipeline.padding(new UnsafeBuffer(in), 0, in.length) >= 0);
     }
 
-    private ProtobufModelHandlerImpl newHandler(
-        String record)
+    private AvroModelHandlerImpl newHandler()
     {
         TestCatalogConfig catalog = CatalogConfig.builder(TestCatalogConfig::new)
             .namespace("test")
             .name("test0")
             .type("test")
             .options(TestCatalogOptionsConfig::builder)
-                .id(1)
+                .id(9)
                 .schema(SCHEMA)
                 .build()
             .build();
-        ProtobufModelConfig model = ProtobufModelConfig.builder()
+        AvroModelConfig model = AvroModelConfig.builder()
             .view("json")
             .catalog()
                 .name("test0")
-                .schema()
-                    .strategy("topic")
-                    .version("latest")
-                    .subject("test-value")
-                    .record(record)
-                    .build()
+                    .schema()
+                        .strategy("topic")
+                        .version("latest")
+                        .subject("test-value")
+                        .build()
                 .build()
             .build();
         when(context.supplyCatalog(catalog.id)).thenReturn(new TestCatalogHandler(catalog.options));
-        return new ProtobufModelHandlerImpl(model, context);
+        return new AvroModelHandlerImpl(config, model, context);
     }
 
     private static byte[] concat(
