+++ runtime/model-json/src/main/java/io/aklivity/zilla/runtime/model/json/internal/JsonModelHandlerImpl.java
@@ -38,8 +38,8 @@ import io.aklivity.zilla.runtime.engine.model.function.ValueConsumer;
 import io.aklivity.zilla.runtime.model.json.config.JsonModelConfig;
 
 // Per-worker factory for a JSON model. One handler serves both directions: supplyDecoder vends a
-// per-stream JsonReadModelPipeline (catalog framing stripped, value validated) and supplyEncoder vends a
-// per-stream JsonWriteModelPipeline (catalog framing emitted, value validated). Configuration-derived
+// per-stream JsonDecodeModelPipeline (catalog framing stripped, value validated) and supplyEncoder vends a
+// per-stream JsonEncodeModelPipeline (catalog framing emitted, value validated). Configuration-derived
 // state (catalog, schema cache, extraction paths) is shared; in-flight state lives on each pipeline.
 public final class JsonModelHandlerImpl extends JsonModelHandler implements ModelHandler
 {
@@ -79,14 +79,14 @@ public final class JsonModelHandlerImpl extends JsonModelHandler implements Mode
     public ModelPipeline supplyDecoder(
         ModelVisitor visitor)
     {
-        return new JsonReadModelPipeline(this, paths, names, visitor);
+        return new JsonDecodeModelPipeline(this, paths, names, visitor);
     }
 
     @Override
     public ModelPipeline supplyEncoder(
         ModelVisitor visitor)
     {
-        return new JsonWriteModelPipeline(this);
+        return new JsonEncodeModelPipeline(this);
     }
 
     int decodePadding(
